+++ perl/LinkedSpec/ActionIR/CanonicalEvents/Core.pm
@@ -35,6 +35,7 @@ my %DIRECT_KIND_BY_CONTRACT_ID = (
  capture                          => 'CAPTURE',
  capture_slice                    => 'CAPTURE_SLICE',
  capture_slice_len                => 'CAPTURE_SLICE_LEN',
+ capture_slice_line               => 'CAPTURE_SLICE_LINE_READ',
  capture_slice_length             => 'CAPTURE_SLICE_LEN',
  start_capture_slice              => 'CAPTURE_SLICE_START',
  capture_slice_here               => 'CAPTURE_SLICE_START',